requirements.txt
speechrecognition
pyttsx3
selenium
wikipedia
wolframalpha
python-dotenv
customtkinter
pillow
packaging
requests
psutil
screen-brightness-control
pycaw
comtypes
google-generativeai
opencv-python
cryptography
pywin32
